requirements.txt
aiofiles==25.1.0
fastapi==0.115.12
openai==1.82.0
psycopg2-binary==2.9.11
python-jose==3.5.0
SQLAlchemy==2.0.41
uvicorn==0.34.3